+++ cybersyn/scripts/mod-compatibility/se-elevator-travel.lua
